+++ R/extend.R
@@ -28,6 +28,9 @@ extend_impl <- function(source, parsed,
   have <- column_names(source)
   check_have_cols(have, partitionby, "rquery::extend partitionby")
   check_have_cols(have, orderby, "rquery::extend orderby")
+  # these checks are easy by the no same block use rule
+  check_have_cols(have, merge_fld(parsed, "symbols_used"), "rquery::extend terms")
+  check_have_cols(have, merge_fld(parsed, "free_symbols"), "rquery::extend terms")
   assignments <- unpack_assignments(source, parsed)
   r <- list(source = list(source),
             table_name = NULL,